validation fix for dog endpoints

+++ src/controllers/dogController.js
@@ -2,7 +2,11 @@ const dogService = require("../services/dogService");
 
 async function getAllDogs(req, res) {
   const allDogs = await dogService.getAllDogs();
-  res.send({ status: "OK", data: allDogs });
+  if (allDogs === undefined || allDogs.length == 0) {
+    res.status(404).send({ status: 404, message: "No record!" });
+  } else {
+    res.status(200).send({ status: "OK", data: allDogs });
+  }
 }
 
 async function getOneDog(req, res) {
@@ -13,7 +17,13 @@ async function getOneDog(req, res) {
     return;
   }
   const dog = await dogService.getOneDog(dogId);
-  res.send({ status: "OK", data: dog });
+  if (dog === undefined || dog.length == 0) {
+    res
+      .status(404)
+      .send({ status: 404, message: "Dog on given id does not exists!" });
+  } else {
+    res.status(200).send({ status: 200, data: dog });
+  }
 }
 
 async function createNewDog(req, res) {
@@ -27,10 +37,12 @@ async function createNewDog(req, res) {
     characteristic: body.characteristic,
   };
   const createdDog = await dogService.createNewDog(newDog);
-  if(!createdDog.hasOwnProperty('error')){
+  if (!createdDog.hasOwnProperty("message")) {
     res.status(201).send({ status: 201, data: createdDog });
-  }else{
-    res.status(createdDog.status).send({ status: createdDog.status, data: createdDog.message });
+  } else {
+    res
+      .status(createdDog.status)
+      .send({ status: createdDog.status, data: createdDog.message });
   }
 }
 
@@ -42,8 +54,24 @@ async function updateOneDog(req, res) {
   if (!dogId) {
     return;
   }
-  const updatedDog = await dogService.updateOneDog(dogId, body);
-  res.send({ status: "OK", data: updatedDog });
+  const dog = await dogService.getOneDog(dogId);
+  if (dog === undefined || dog.length == 0) {
+    res
+      .status(404)
+      .send({
+        status: 404,
+        message: "Could not update dog, given id does not exists!",
+      });
+  } else {
+    const updatedDog = await dogService.updateOneDog(dogId, body);
+    if (!updatedDog.hasOwnProperty("message")) {
+      res.status(200).send({ status: "OK", data: updatedDog });
+    } else {
+      res
+        .status(updatedDog.status)
+        .send({ status: updatedDog.status, data: updatedDog.message });
+    }
+  }
 }
 
 async function deleteOneDog(req, res) {
@@ -53,8 +81,18 @@ async function deleteOneDog(req, res) {
   if (!dogId) {
     return;
   }
-  await dogService.deleteOneDog(dogId);
-  res.status(204).send({ status: "OK" });
+  const dog = await dogService.getOneDog(dogId);
+  if (dog === undefined || dog.length == 0) {
+    res
+      .status(404)
+      .send({
+        status: 404,
+        message: "Could not delete dog, given id does not exists!",
+      });
+  } else {
+    await dogService.deleteOneDog(dogId);
+    res.status(200).send({ status: 200, message: "Dog deleted!" });
+  }
 }
 
 module.exports = {
